+++ node/runtime/common/src/impls.rs
@@ -196,3 +196,95 @@ where
 		Ok(())
 	}
 }
+
+#[macro_export]
+macro_rules! impl_self_contained_call {
+	() => {
+		impl fp_self_contained::SelfContainedCall for Call {
+			type SignedInfo = H160;
+
+			fn is_self_contained(&self) -> bool {
+				match self {
+					Call::Ethereum(call) => call.is_self_contained(),
+					_ => false,
+				}
+			}
+
+			fn check_self_contained(
+				&self,
+			) -> Option<Result<Self::SignedInfo, TransactionValidityError>> {
+				match self {
+					Call::Ethereum(call) => call.check_self_contained(),
+					_ => None,
+				}
+			}
+
+			fn validate_self_contained(
+				&self,
+				info: &Self::SignedInfo,
+			) -> Option<TransactionValidity> {
+				match self {
+					Call::Ethereum(call) => call.validate_self_contained(info),
+					_ => None,
+				}
+			}
+
+			fn pre_dispatch_self_contained(
+				&self,
+				info: &Self::SignedInfo,
+			) -> Option<Result<(), TransactionValidityError>> {
+				match self {
+					Call::Ethereum(call) => call.pre_dispatch_self_contained(info),
+					_ => None,
+				}
+			}
+
+			fn apply_self_contained(
+				self,
+				info: Self::SignedInfo,
+			) -> Option<sp_runtime::DispatchResultWithInfo<PostDispatchInfoOf<Self>>> {
+				match self {
+					call @ Call::Ethereum(darwinia_ethereum::Call::transact { .. }) => {
+						Some(call.dispatch(Origin::from(
+							darwinia_ethereum::RawOrigin::EthereumTransaction(info),
+						)))
+					}
+					_ => None,
+				}
+			}
+		}
+
+		fn validate_self_contained_inner(
+			call: &Call,
+			eth_call: &darwinia_ethereum::Call<Runtime>,
+			signed_info: &<Call as fp_self_contained::SelfContainedCall>::SignedInfo,
+		) -> TransactionValidity {
+			if let darwinia_ethereum::Call::transact { ref transaction } = eth_call {
+				// Previously, ethereum transactions were contained in an unsigned
+				// extrinsic, we now use a new form of dedicated extrinsic defined by
+				// frontier, but to keep the same behavior as before, we must perform
+				// the controls that were performed on the unsigned extrinsic.
+				use sp_runtime::traits::SignedExtension as _;
+				let input_len = match transaction {
+					darwinia_ethereum::Transaction::Legacy(t) => t.input.len(),
+					darwinia_ethereum::Transaction::EIP2930(t) => t.input.len(),
+					darwinia_ethereum::Transaction::EIP1559(t) => t.input.len(),
+				};
+				let extra_validation =
+					SignedExtra::validate_unsigned(call, &call.get_dispatch_info(), input_len)?;
+				// Then, do the controls defined by the ethereum pallet.
+				use fp_self_contained::SelfContainedCall as _;
+				let self_contained_validation =
+					eth_call.validate_self_contained(signed_info).ok_or(
+						TransactionValidityError::Invalid(InvalidTransaction::BadProof),
+					)??;
+
+				Ok(extra_validation.combine_with(self_contained_validation))
+			} else {
+				Err(TransactionValidityError::Unknown(
+					sp_runtime::transaction_validity::UnknownTransaction::CannotLookup,
+				))
+			}
+		}
+	};
+}
