+++ node/runtime/common/src/lib.rs
@@ -134,9 +134,5 @@ frame_support::parameter_types! {
 		.get(DispatchClass::Normal);
 }
 
-// The maxinum block weight for normal transaction is MAXIMUM_BLOCK_WEIGHT(2s) * NORMAL_DISPATCH_RATIO = 1_000_000_000_000 * 75% = 750_000_000_000
 // Accounding to the evm gas benchmark, 1 gas ~= 40_000 weight.
 pub const WEIGHT_PER_GAS: u64 = 40_000;
-// pub const BLOCK_GAS_LIMIT: u64 = MAXIMUM_BLOCK_WEIGHT * NORMAL_DISPATCH_RATIO / WEIGHT_PER_GAS;
-pub const EXTRINSIC_BASE_GAS: u64 = ExtrinsicBaseWeight::get() / WEIGHT_PER_GAS;
-// pub const EXTRINSIC_GAS_LIMIT: u64 = BLOCK_GAS_LIMIT - BLOCK_GAS_LIMIT * 0.025 - EXTRINSIC_BASE_GAS;
