+++ tests/dvm/darwinia/tests/config.ts
@@ -8,4 +8,4 @@ export const config = {
 };
 
 // EXTRINSIC_GAS_LIMIT = [BLOCK_GAS_LIMIT - BLOCK_GAS_LIMIT * (NORMAL_DISPATCH_RATIO - AVERAGE_ON_INITIALIZE_RATIO) - EXTRINSIC_BASE_Weight] / WEIGHT_PER_GAS  = (1_000_000_000_000 * 2 * (0.75-0.025) - 125_000_000) / 40_000
-export const EXTRINSIC_GAS_LIMIT = 36246875
+export const EXTRINSIC_GAS_LIMIT = 36246875;
